% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_summary.R
\name{total_summary}
\alias{total_summary}
\title{Create a Detailed Summary Table for a Variable}
\usage{
total_summary(data, ..., summ_var, na.rm = FALSE)
}
\arguments{
\item{data}{The dataframe to be used.}

\item{...}{The variable(s) to group the summary variable by. It is not necessary to group the summary variable.}

\item{summ_var}{The variable for which a summary table will be created.}

\item{na.rm}{Set to TRUE to remove NA values from data processed by the function.}
}
\value{
A table
}
\description{
Creates a table of summary statistics for a variable. These summary statistics are the mean, range, minimum, first quartile, median, third quartile, and maximum values of the variable.
}
\examples{
data <- data.frame(
  colour = rep(c("Red", "Blue", "Yellow"), each = 5),
  number = rnorm(15))
total_summary(data, colour, summ_var = number)
}
